DuAn1_PRO1041/test/Repositories/Impl/KhachHangRepositoryImplTest.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/UnitTests/JUnit4TestClass.java to edit this template
 */
package Repositories.Impl;

import DomainModel.KhachHang;
import java.util.List;
import org.junit.After;
import org.junit.AfterClass;
import org.junit.Before;
import org.junit.BeforeClass;
import org.junit.Test;
import static org.junit.Assert.*;

/**
 *
 * @author Chien Duong
 */
public class KhachHangRepositoryImplTest {
    
    public KhachHangRepositoryImplTest() {
    }
    
    @BeforeClass
    public static void setUpClass() {
        // Chuẩn bị các tài nguyên toàn cục trước khi chạy tất cả các test case
    }
    
    @AfterClass
    public static void tearDownClass() {
        // Giải phóng tài nguyên toàn cục sau khi chạy tất cả các test case
    }
    
    @Before
    public void setUp() {
        // Chuẩn bị tài nguyên cần thiết trước khi chạy mỗi test case
    }
    
    @After
    public void tearDown() {
        // Giải phóng tài nguyên sau khi chạy mỗi test case
    }

    /**
     * Test of getAll method, of class KhachHangRepositoryImpl.
     */
    @Test
    public void testGetAll() {
        System.out.println("getAll");
        KhachHangRepositoryImpl instance = new KhachHangRepositoryImpl();
        // Gọi phương thức getAll để lấy danh sách khách hàng
        List<KhachHang> result = instance.getAll();
        // Kiểm tra xem kết quả có phải là null không
        assertNotNull("Danh sách khách hàng không được là null", result);
        // TODO: Thêm các kiểm tra khác nếu cần thiết
        
        // Không sử dụng fail() vì chúng ta đã có kiểm tra assertNotNull ở trên
    }

    /**
     * Test of getListMa method, of class KhachHangRepositoryImpl.
     */
    @Test
    public void testGetListMa() {
        System.out.println("getListMa");
        KhachHangRepositoryImpl instance = new KhachHangRepositoryImpl();
        // Gọi phương thức getListMa để lấy danh sách mã khách hàng
        List<String> result = instance.getListMa();
        // Kiểm tra xem kết quả có phải là null không
        assertNotNull("Danh sách mã khách hàng không được là null", result);
        // TODO: Thêm các kiểm tra khác nếu cần thiết
        
        // Không sử dụng fail() vì chúng ta đã có kiểm tra assertNotNull ở trên
    }

    /**
     * Test of add method, of class KhachHangRepositoryImpl.
     */
    @Test
    public void testAdd() {
        System.out.println("add");
        KhachHang kh = new KhachHang(); // Tạo đối tượng KhachHang mới để thêm vào repository
        KhachHangRepositoryImpl instance = new KhachHangRepositoryImpl();
        // Gọi phương thức add để thêm khách hàng
        boolean result = instance.add(kh);
        // Kiểm tra xem kết quả có phải là true không
        assertTrue("Thêm khách hàng không thành công", result);
        // TODO: Thêm các kiểm tra khác nếu cần thiết
        
        // Không sử dụng fail() vì chúng ta đã có kiểm tra assertTrue ở trên
    }

    /**
     * Test of update method, of class KhachHangRepositoryImpl.
     */
    @Test
    public void testUpdate() {
        System.out.println("update");
        KhachHang kh = new KhachHang(); // Tạo đối tượng KhachHang mới để cập nhật
        String id = "123"; // Giả sử id của khách hàng cần cập nhật là "123"
        KhachHangRepositoryImpl instance = new KhachHangRepositoryImpl();
        // Gọi phương thức update để cập nhật khách hàng
        boolean result = instance.update(kh, id);
        // Kiểm tra xem kết quả có phải là true không
        assertTrue("Cập nhật khách hàng không thành công", result);
        // TODO: Thêm các kiểm tra khác nếu cần thiết
        
        // Không sử dụng fail() vì chúng ta đã có kiểm tra assertTrue ở trên
    }

    /**
     * Test of delete method, of class KhachHangRepositoryImpl.
     */
    @Test
    public void testDelete() {
        System.out.println("delete");
        String id = "123"; // Giả sử id của khách hàng cần xóa là "123"
        KhachHangRepositoryImpl instance = new KhachHangRepositoryImpl();
        // Gọi phương thức delete để xóa khách hàng
        boolean result = instance.delete(id);
        // Kiểm tra xem kết quả có phải là true không
        assertTrue("Xóa khách hàng không thành công", result);
        // TODO: Thêm các kiểm tra khác nếu cần thiết
        
        // Không sử dụng fail() vì chúng ta đã có kiểm tra assertTrue ở trên
    }

    /**
     * Test of addNhanh method, of class KhachHangRepositoryImpl.
     */
    @Test
    public void testAddNhanh() {
        System.out.println("addNhanh");
        KhachHang kh = new KhachHang(); // Tạo đối tượng KhachHang mới để thêm nhanh vào repository
        KhachHangRepositoryImpl instance = new KhachHangRepositoryImpl();
        // Gọi phương thức addNhanh để thêm nhanh khách hàng
        boolean result = instance.addNhanh(kh);
        // Kiểm tra xem kết quả có phải là true không
        assertTrue("Thêm nhanh khách hàng không thành công", result);
        // TODO: Thêm các kiểm tra khác nếu cần thiết
        
        // Không sử dụng fail() vì chúng ta đã có kiểm tra assertTrue ở trên
    }
}
